Optimize router delete execution
Optimize the execution of router delete in the L3 agent
by not performing operations in the router namespace that
will be removed as the last step of the workflow.
Comparing the slopes of the least squares fit of delete
execution time shows that this optimization decreases that
slope by over half.

Change-Id: Ic425ab1f8c3afe882c9deaa72a814ac0d2053f3b
Signed-off-by: Ryan Moats <rmoats@us.ibm.com>
Closes-Bug: 1512450

Conflicts:
  neutron/agent/l3/dvr_edge_ha_router.py
  neutron/agent/l3/dvr_edge_router.py
  neutron/agent/l3/dvr_router.py
  neutron/agent/l3/dvr_router_base.py
  neutron/agent/l3/router_info.py