Check if namespace exists before exec commands
If a neutron router has no ports defined, its namespace is deleted.
It is required to check if the namespace exists before calling
exec commands against the namespace.
It is not enough for the string of the namespace uuid to be defined,
the namespace must really exist on the network node.
Otherwise the agent log file will be flooded with messages like
Cannot open network namespace "qrouter-<uuid>": No such file or directory
Failed to process compatible router '<uuid>'

Related-bug: 1573073

Change-Id: I744ef11529f9da5cbfdb812de0b35b95f9d078bb